services/ingestion/requirements.txt
pillow
click
pymongo
pandas
pdfminer.six
chardet
opencv-python
camelot-py[cv]
